+++ CHANGES.md
@@ -5,6 +5,7 @@
 
 ## Fixes
 * Fixed loading of game configuration settings
+* Fixed game setting screen
 * Fixed loading of trait properties
 * Fixed House Pet / FO training names
 * Fixed loading of special stats